+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - Add support for `kebab-case` executable names. [#205](https://github.com/PyO3/setuptools-rust/pull/205)
+- Add support for custom cargo profiles. [#216](https://github.com/PyO3/setuptools-rust/pull/216)
 
 ## 1.1.2 (2021-12-05)
 ### Changed